+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 * extend **SparqlSourceProcessor**: add retries on failures configurable with parameters `chunkSizeDecreaseFactor` and `maxRetries`
 * extend **SparqlSourceProcessor**: add parameter `followInverseUnlimited`
+* extend **LiteralValueComparisonProcessor**: add parameter `languageFilterPatterns`
 
 ### Fixed
 * fix **EquivalentValueMappingProcessor**: fix message format